new debuggong scene delegate to handle debugging scenario, thus make the default scene clean from any test logic
