+++ FeedApp/FeedApp/SceneDelegate.swift
@@ -13,8 +13,11 @@ import CoreData
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
     var window: UIWindow?
-
-
+    
+    let localStoreURL = NSPersistentContainer
+                            .defaultDirectoryURL()
+                            .appendingPathComponent("FeedStore")
+    
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let _ = (scene as? UIWindowScene) else { return }
         
@@ -23,16 +26,10 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         let remoteClient = makeRemoteClient()
         let remoteImageLoader = RemoteFeedImageDataLoader(client: remoteClient)
         let remoteFeedLoader = RemoteFeedLoader(client: remoteClient, url: url)
-        
-        let localStoreURL = NSPersistentContainer
-                                .defaultDirectoryURL()
-                                .appendingPathComponent("FeedStore")
-        
-        #if DEBUG
+                
         if CommandLine.arguments.contains("-reset") {
             try? FileManager.default.removeItem(at: localStoreURL)
         }
-        #endif
         
 //        let localStore = try! CoreDataFeedStore(storeURL: localStoreURL)
 //        let localFeedLoader = LocalFeedLoader(store: localStore, currentDate: Date.init)
@@ -56,26 +53,8 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         window?.rootViewController = backupfeedVC
     }
     
-    private func makeRemoteClient() -> HTTPClient {
-        #if DEBUG
-        if UserDefaults.standard.string(forKey: "connectivity") == "offline" {
-            return AlwaysFailingHTTPClient()
-        }
-        #endif
+    func makeRemoteClient() -> HTTPClient {
         return URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
     }
 
 }
-
-#if DEBUG
-private final class AlwaysFailingHTTPClient: HTTPClient {
-    private class Task: HTTPClientTask {
-        func cancel() { }
-    }
-    
-    func get(from url: URL, completion: @escaping (HTTPResponse) -> Void) -> HTTPClientTask {
-        completion(.failure(NSError(domain: "any error", code: 0)))
-        return Task()
-    }
-}
-#endif
